Dużo ciekawych rzeczy ;] - konwerter list -> std::set (+dopisek do dokumentacji) - usunięte warningi signed/unsigned poprzez rzutowanie (int) - dostosowanie nazw bibliotek do sconsa - zmienione nazwy TestModule - na Foo, Foobar - drobne poprawki w External

+++ Source/SetWrapper_py_impl.hpp
@@ -56,7 +56,7 @@ void SetWrapper<V>::print(const Set& x) {
     }
     typename std::set<V>::const_iterator it = x.begin();
     std::cout << "set([";
-	for(int i = 0; i < x.size() - 1; i++)
+	for(int i = 0; i < (int)x.size() - 1; i++)
 	{
         std::cout << *it << ", ";
         ++it;
